requirements.txt
click>=6.7
colorama>=0.3.3
cryptography>=2.0.3
requests>=2.4.3
PyYAML>=3.10
pyvcloud>=18.0.15.dev4
tabulate>=0.7.5
python-dateutil>=2.4.2
xmltodict>=0.9.2
Pygments
PrettyTable>=0.7,<0.8
six
pytz
oslo.utils
pycrypto
keyring
